Add force RFC2965 cookie decoder option
Current ServerCookieDecoder's implementation satisfies RFC2965 and RFC6265 specification.
There, however, might be a case that some clients using Android(or OpenJDK) `HttpCookie` and `CookieManager` don't include "$VERSION".

Details are below.

- "$Version" is set only if minimum version header is `1`.[1]
- The version header is decided in here (by heuristic) and it maybe 0.[2]
- This behavior is for compatibility with RFC2109 and RFC2965.[3]

This is a problem of android implementation, but some android applications are very difficult to update..., so we want to have an escape hatch config which enforces parsing cookies with RFC2965 style (whether "$VERSION" field is absent or not).

[1]: https://android.googlesource.com/platform/libcore/+/android-4.3_r3/luni/src/main/java/java/net/CookieManager.java#130
[2]: https://android.googlesource.com/platform/libcore/+/android-4.3_r3/luni/src/main/java/java/net/HttpCookie.java#217
[3]: https://android.googlesource.com/platform/libcore/+/android-4.3_r3/luni/src/main/java/java/net/HttpURLConnection.java#205